Move specific APIErrors to their own module
We may want to use this from e.g. haas.model, which would create
circular dependencies. This way we don't have to worry about that.
